chore: added changes for registration purpose
