+++ bc_obps/reporting/service/report_verification_service.py
@@ -6,9 +6,9 @@ from reporting.models.report_verification_visit import ReportVerificationVisit
 from reporting.models import ReportVersion
 
 from registration.models import Operation
-from reporting.service.report_additional_data import ReportAdditionalDataService
 from reporting.service.compliance_service import ComplianceService
 from reporting.schema.report_verification import ReportVerificationIn
+from service.report_service import ReportService
 
 
 class ReportVerificationService:
@@ -97,7 +97,7 @@ class ReportVerificationService:
         ATTRIBUTABLE_EMISSION_THRESHOLD = Decimal("25000")  # 25,000 TCo₂e
 
         # Fetch registration purpose
-        registration_purpose = ReportAdditionalDataService.get_registration_purpose_by_version_id(version_id)
+        registration_purpose = ReportService.get_registration_purpose_by_version_id(version_id)
         registration_purpose_value = registration_purpose.get("registration_purpose", {})
 
         # Check the REGULATED_OPERATION_PURPOSES
