+++ bc_obps/reporting/tests/service/test_report_additional_data_service.py
@@ -73,14 +73,3 @@ class TestReportAdditionalDataService(TestCase):
         self.assertEqual(report_additional_data.emissions_off_site_transfer, data.emissions_off_site_transfer)
         self.assertEqual(report_additional_data.electricity_generated, data.electricity_generated)
         self.assertEqual(report_additional_data.report_version, self.report_version)
-
-    def test_get_registration_purpose_by_version_id_returns_correct_data(self):
-        """
-        Test that the service retrieves the correct registration purpose
-        for a given report version ID.
-        """
-        retrieved_data = ReportAdditionalDataService.get_registration_purpose_by_version_id(
-            version_id=self.report_version.id
-        )
-        self.assertIsNotNone(retrieved_data)
-        self.assertEqual(retrieved_data["registration_purpose"], self.report_operation.registration_purpose)
